Moves redux middleware to actions/utils
