+++ src/actions/ast.js
@@ -10,7 +10,7 @@ import {
 } from "../selectors";
 
 import { getMappedExpression } from "./expressions";
-import { PROMISE } from "../utils/redux/middleware/promise";
+import { PROMISE } from "./utils/promise";
 import {
   getSymbols,
   getEmptyLines,
