+++ src/actions/pause/commands.js
@@ -2,7 +2,7 @@
 
 import { getPause, getSelectedSource } from "../../selectors";
 import { getPausedPosition } from "../../utils/pause";
-import { PROMISE } from "../../utils/redux/middleware/promise";
+import { PROMISE } from "../utils/promise";
 import { getNextStep } from "../../workers/parser";
 import { addHiddenBreakpoint } from "../breakpoints";
 import { features } from "../../utils/prefs";
