+++ src/actions/utils/history.js
@@ -1,8 +1,13 @@
 // @flow
 
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+/* global window */
+
 import { isDevelopment } from "devtools-config";
 
-import type { ThunkArgs } from "../../../actions/types";
+import type { ThunkArgs } from "../types";
 
 /**
  * A middleware that stores every action coming through the store in the passed
