+++ src/actions/utils/thunk.js
@@ -1,6 +1,11 @@
 // @flow
 
-import type { ThunkArgs, ActionType } from "../../../actions/types";
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+/* global window */
+
+import type { ThunkArgs, ActionType } from "../types";
 
 /**
  * A middleware that allows thunks (functions) to be dispatched. If
