+++ src/actions/utils/timing.js
@@ -1,3 +1,8 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+/* global window */
+
 /**
  * Redux middleware that sets performance markers for all actions such that they
  * will appear in performance tooling under the User Timing API
